Major refactoring of the ParametricOctree 2D/3D models such that now the generation of the dummy geometry information for the grid/topology of the octree_dmodel relies on functions written in GridapP4est, reducing code replication.
Also both the 2D and the 3D parametric models come with boundary
face labels. In the case of the 2D model, all vertices, edges, faces,
cells are labeled with the "boundary" tag.

In the case of the 3D model, we have three different tags,
"bottom_boudary", "top_boundary" and "intermediate_boundary".
For the cubed sphere with thickness, the latter name is a bit
misleading as it actually includes vertices, edges, faces, cells
which are interior to the domain.

This work relies on a branch in GridapP4est, namely:

https://github.com/gridap/GridapP4est.jl/tree/remove_dependence_on_interior_boundary_tag_coarse_model_p6est

Once all that work becomes more stable, we will release the changes in
this branch in a new version GridapP4est 0.13.12